code clean

+++ src/Geoscape/BaseDefenseState.cpp
@@ -162,7 +162,7 @@ void BaseDefenseState::btnOkClick(Action *)
 		_game->popState();
 		if (_base->getSoldiers())
 		{
-			int month = _game->getSavedGame()->getMonthsPassed();
+			size_t month = _game->getSavedGame()->getMonthsPassed();
 			if (month > _game->getRuleset()->getAlienItemLevels().size()-1)
 				month = _game->getRuleset()->getAlienItemLevels().size()-1;
 			SavedBattleGame *bgame = new SavedBattleGame();
